Set default fd limit to 4096
An earlier problem that resulted in excessive fd allocation during
restoration of a backup consisting of a very large number of files
would've caused the default 1024 fd limit to be exceeded for not-
so-terribly-large database. While that problem has been fixed, the
default limit is still fairly low and large systems may end up
doing downloads on a very large number of threads and the storage
provider specific SDKs may allocate unknown number of fds per
transfer. Bump the limit up to 4096 to avoid running out of file
descriptors under normal circumstances.