Remove excessive bytes slicing from decryption
Perf testing shows that restoring backups is slow because a lot of time
is spent doing memmove. The decryption code does a whole lot of slicing
and joins for bytes and might contribute to that. Test an alternative
implementation that avoids slicing as much as possible with the cost of
not doing memory buffering. If dropping memory buffering turns out to
add a lot of IO load it could be brought back in somehow more finetuned
manner so that it doesn't require as much slicing/joining.

This implementation also fixes major issue with seek to end-of-file
handling, which previously read and decrypted full file contents into
memory and then threw the results away. Seeking to eof is common for
getting size of file, which we can get practically as a zero op.

Note that this implementation only works as long as we're using AES-CTR
since it makes assumptions regarding unencrypted data that may not hold
for other ciphers / cipher modes.